fix(db): align JPA entity mappings with database enum constraints
- replace @JdbcTypeCode(OTHER) with @ColumnTransformer for JSONB columns to fix PostgreSQL compatibility
- change lengthSort type from Double to BigDecimal for sorting precision
- migrate sys_user status and auth_refresh_token revoke_reason to Java enum values
- add check constraints matching @Enumerated(EnumType.STRING) definitions

+++ src/main/java/com/leo/erp/master/material/domain/entity/Material.java
@@ -59,7 +59,7 @@ public class Material extends AuditableEntity {
     private String remark;
 
     @Column(name = "length_sort", insertable = false, updatable = false)
-    private Double lengthSort;
+    private BigDecimal lengthSort;
 
     @Column(name = "spec_sort", insertable = false, updatable = false)
     private Integer specSort;
